Move to Python 3 to support tar/zip files using .xz
In Python 3, the standard library tarfile and zipfile modules
automatically support archives using .xz/LZMA compression. We
now want this and switching to Python 3 is reasonably easy.

This is only an initial version. It appears to work, but it's
probably doing some things (like logging) in a backwards way
and it hasn't been fully hardened against Unicode decoding and
encoding errors.