Attempt to catch exceptions for some ZIP problems
The zipfile module signals certain errors by raising standard Python
interpreter exceptions, including 'this ZIP archive is encrypted and
either you didn't give me a password or it's wrong'. Password problems
raise RuntimeError; certain other issues raise NotImplementedError.
This is extremely noxious behavior, of course, but what can we do.  Now
we catch both exceptions when we call .read() on ZIP contents as part of
trying to scan zips-in-a-zip.

Scanning zips-in-zips is the only time encryption appears to matter for
our current use. File names are not stored encrypted, only file content,
so we can still get file type information for an attachment that is an
encrypted ZIP archive.