Potentially catch and syslog surprise exceptions
It's always possible for us to hit a surprise exception (perhaps because
file examination modules return undocumented exceptions for some
situations, for example the zipfile module and encrypted ZIP archives).
In our usual configuration for running alogger.py, nothing notices or
logs the resulting failure and things just fail silently. Silent failure
is bad, especially when it's a sign that we don't understand what
exceptions a module we're using can actually raise, so now we can catch
and syslog any surprise exceptions.

This is only done if syslog output is enabled in general. If you haven't
enabled it, the surprise exception is simply propagated outwards (where
it will produce the usual Python stack trace to standard error).

(Perhaps this is a hack, but it's a convenient one.)