Improve the handling of non-ASCII in logged messages
Our previous method of making messages safe to log (either to standard
output or to syslog) was rather crude. Now we do this in a better way
that more clearly tracks what we want to escape and how, and in the
process we now write (carefully escaped) Unicode strings to stdout and
syslog instead of handing bytes to the log functions and having them do
weird things.

(In the process I discovered that syslog always encodes its messages in
UTF-8, probably along with a bunch of other standard modules that use
the same C interfaces. This may be documented somewhere general.)