Deal with a missing Content-Disposition header
In theory all MIME parts have a Content-Disposition. In practice,
malware sends us attachments without C-D headers, and we were previously
assuming that this only happened if we were processing the body of a
non-multipart message. We were also assuming that such a body is always
uninteresting, which in practice is not necessarily the case; someone
might send, say, a ZIP file as the body.

Now we no longer assume that a missing C-D value makes the part
automatically uninteresting. Instead we treat it much the same as a C-D
of 'inline', which gets at least some cases.

(A missing C-D on an actual MIME part is a danger sign and should
probably always provoke a report, but right now we can't tell if we're
dealing with a MIME part or the main message body of a non-multipart
message.)