Handle zero-length filenames from ZIP archives
Under some circumstances the zipfile module can give us filenames that
are empty; this happens when some clever person puts a null byte as the
first byte of the raw in-ZIP filename, because zipfile truncates the
filenames it gives back at the first null byte. At the moment we opt to
deal with this rather than fish out the pre-truncation filename.