Get some additional file type information from libmagic
If we don't have a file extension, we now attempt to get some
information from libmagic's file type sniffing via the third-party
'magic' module. We also apply this to file extensions that are
just (some) compression formats like '.gz'; since libmagic can't
currently sniff inside .7z files, we leave those alone.

If the magic-determined file type information is the same as the
declared MIME type, we don't bother reporting it.

(We could sniff more things in the interests of cross-checking declared
file extensions against the detected reality, but right now we're not
interested in going that far. This is a backup for additional
information if we're lacking it, not really a cross-check to pick up
lies in information that's under an attacker's control.)