Handle sniffing file type for compressed files with extensions
We were already using libmagic to sniff file types for compressed
files, eg 'file.gz', but due to an oversight we weren't doing this
when the MIME filename had an additional extension, eg
'file.txt.gz'. This happened because the extension code reports the
extension as '.txt.gz' instead of just '.gz'. Now we deal with this
correctly so that we will sniff file type information for malware that
shows up as eg '.7283nsn.gz'.