Don't bother saying a real .tar.gz is an application/x-tar file
After the previous change, we would use libmagic to sniff the content
type of a '.tar.gz' file even if it really was a genuine tar file that
we could understand fine. This is both redundant and relatively
common, so now we catch this case and don't report on it.