Try to sniff libmagic info about a .jar that isn't a ZIP archive
If we see a (MIME) filename extension of .jar but our 'is this a ZIP
archive' check says 'no', we now check via libmagic to see what it may
actually be. This is the same check as we were already doing for .zip
extensions, just extended to .jar ones.

(Some such MIME attachments have been detected in the wild.)