Report specifically when an archive only has directories
When we inspect the contents of an archive (in any format), we
normally only report on the extensions of files, if there are any; if
there are no files, we report that. With this change, we now
distinguish between the case of an archive with no contents at all and
an archive that only contains directories (that we wouldn't otherwise
report about).