Work around tarfile mis-detecting things as tar files
The standard library tarfile module has a problem where under some
circumstances it will mis-detect files as empty tar files when they
very much aren't. The known circumstance is a file that starts with
512 bytes of zero bytes. We work around this using internal knowledge
of how the tarfile module behaves, so that we can immediately detect
this case. This workaround is a hack but it beats our other options.

Python issue: https://bugs.python.org/issue36596