Add support for looking inside archives with libarchive
The C libarchive library allows us to look inside a number of archive
types that are otherwise not accessible, including 7zip and ISO
images. We now support using it in general, and attempt to use it on
7zip, ISO 9660, CAB, and LHA archives. Unfortunately libarchive has no
general 'try to open this and tell me the format', so we try various
specific formats.