Deal with libarchive being unable to provide a filename to us
Under some circumstances libarchive can be unable to provide a
filename to us (because of character set encoding issues) and returns
a None for the filename instead of a string. Detect and skip these
filenames. As a defensive measure, also detect the general issue in
the code that processes filename lists.